refactor(frontend): Rename component of License Agreement link
