+++ src/frontend/src/lib/components/license-agreement/LicenseAgreementLink.svelte
